Addressed adams comment
