--- acceptance_tests/utilities/support_tool_healthcheck.py
@@ -1,44 +1,0 @@
-import argparse
-from time import sleep
-
-from requests.exceptions import ConnectionError, HTTPError
-
-from acceptance_tests.utilities.iap_requests import make_request
-
-
-def support_tool_healthcheck(url: str, max_retries: int):
-    check_support_tool = True
-    num_of_retries = 0
-    while check_support_tool:
-        try:
-            response = make_request(url=url)
-            response.raise_for_status()
-            check_support_tool = False
-        except (ConnectionError, HTTPError):
-            num_of_retries += 1
-            if num_of_retries == max_retries:
-                print("Reached maximum number of retries. Stopping script")
-                exit(1)
-
-            print("Support Tool not available. Sleeping for a minute")
-            sleep(60)
-
-
-def parse_arguments():
-    parser = argparse.ArgumentParser(
-        description="This script polls the Support Tool health endpoint to check if it's available for the Acceptance "
-                    "Tests to start running")
-    parser.add_argument('--url', help='The Support Tool url to pass in', type=str, required=True)
-    parser.add_argument('--max_retries', help='Maximum number of retries we want to check support tool is up', type=int,
-                        required=True)
-
-    return parser.parse_args()
-
-
-def main():
-    args = parse_arguments()
-    support_tool_healthcheck(url=args.url, max_retries=args.max_retries)
-
-
-if __name__ == "__main__":
-    main()
