+++ run_gke.sh
@@ -48,7 +48,7 @@ kubectl apply -f $COMPLETE_MANIFEST
 kubectl wait --for=condition=Ready pod/acceptance-tests --timeout=200s
 
 kubectl exec -it acceptance-tests -- /bin/bash -c \
-  "sleep 2;  python -m acceptance_tests.utilities.support_tool_healthcheck --url https://support-tool-$ENV.rm.gcp.onsdigital.uk/actuator/health --max_retries 10"
+  "sleep 2;  python -m poll_endpoint --url https://support-tool-$ENV.rm.gcp.onsdigital.uk/actuator/health --max_retries 10"
 
 
 kubectl exec -it acceptance-tests -- /bin/bash -c "sleep 2; behave acceptance_tests/features $BEHAVE_TAGS --logging-level WARN"
